backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
google-generativeai==0.3.2
Pillow==10.1.0
django-filter==23.5 